+++ src/namdesktop/model/NodeStatus.java
@@ -1,0 +1,5 @@
+package namdesktop.model;
+
+public enum NodeStatus {
+    ACTIVE, DONE, CANCELLED, ARCHIVED
+}